requirements.txt
streamlit
requests
beautifulsoup4
lxml
pyyaml
pillow
playwright
deep-translator
